// registration/src/Components/Models/User.ts
export type User = {
  f_name: string;
  l_name: string;
  NPI: number;
  address: {
    address1: string;
    address2: string;
    city: string;
    state: string;
    zip: number;
  };
  phone_num: number;
  email: string;
};
// •	First Name
// •	Last Name
// •	NPI number
// •	Business Address
// •	Telephone Number
// •	Email address